pyproject.toml
[project]
name = "graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "sqlalchemy>=2.0.40",
    "strawberry-graphql[debug-server]>=0.263.2",
]
